added database schema

+++ src/main/java/com/alexecollins/openbookmaker/sports/model/Price.java
@@ -15,5 +15,5 @@ public class Price implements Serializable {
 	}
 
 	@NonNull
-	private short num, den;
+	private int num, den;
 }
